Tests/LibWeb/Text/expected/css/viewport-overflow-hidden-remains-programmatically-scrollable.txt

Overflow extends scroll width: true
Horizontal scroll offset: 50
